Add confirmation before deleting

See merge request chas-challenge-2026/grupp-13/burgundy!58

+++ frontend/Volunti.Web/src/components/organization/OrgDashboard.jsx
@@ -140,6 +140,9 @@ export default function OrgDashboard() {
   };
 
   const handleDeleteJob = async (jobId) => {
+    if (!window.confirm("Är du säker på att du vill ta bort jobbet?")){
+      return;
+    }
     try {
       await deleteJob(jobId);
       await loadDashboardData();
